requirements.txt
torch==2.7.1
torchvision==0.22.1
tqdm==4.67.1
matplotlib==3.10.3
ipykernel==6.29.5
wandb==0.21.0
python-dotenv==1.1.1
hydra-core==1.3.2
einops==0.8.1
scipy==1.16.1
pytz==2025.2
datasets==4.0.0